# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  response.BadRequestResponse:
    description: Bad Request response format
    properties:
      code:
        example: 4000
        type: integer
      message:
        example: Invalid request parameters
        type: string
    type: object
...
  response.ForbiddenResponse:
    description: Forbidden response format
    properties:
      code:
        example: 4003
        type: integer
      message:
        example: Access denied
        type: string
    type: object
  response.InternalServerErrorResponse:
    description: Internal Server Error response format
    properties:
      code:
        example: 5000
        type: integer
      message:
        example: Internal server error
        type: string
    type: object
...
  response.NotFoundResponse:
    description: Not Found response format
    properties:
      code:
        example: 4004
        type: integer
      message:
        example: Resource not found
        type: string
    type: object
  response.PaginatedResponse:
    description: Paginated response format
    properties:
      code:
        example: 0
        type: integer
      data:
        items:
          type: object
        type: array
      limit:
        example: 10
        type: integer
      message:
        example: success
        type: string
      offset:
        example: 0
        type: integer
      total:
        example: 100
        type: integer
    type: object
...
  response.UnauthorizedResponse:
    description: Unauthorized response format
    properties:
      code:
        example: 4001
        type: integer
      message:
        example: Authentication required
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: A comprehensive service management platform with subscription-based
    billing, user management, and server administration. Features include OAuth2 authentication,
    traffic subscription management, multi-gateway payments, referral programs, and
    customer support system.
  title: Linke API
  version: "1.0"
paths:
...
  /admin/payments/configs:
    get:
      consumes:
      - application/json
      description: Get payment configurations with filtering and pagination
      parameters:
      - description: Filter by gateway
        example: '"epay"'
        in: query
        name: gateway
        type: string
      - description: Filter by method
        example: '"alipay"'
        in: query
        name: method
        type: string
      - description: Filter by enabled status
        example: true
        in: query
        name: is_enabled
        type: boolean
      - description: Filter by environment
        example: '"production"'
        in: query
        name: environment
        type: string
      - description: Limit results
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Offset results
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PaymentConfigResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Get payment configs'
      tags:
      - Admin-Payment-Management
    post:
      consumes:
      - application/json
      description: Create a new payment configuration
      parameters:
      - description: Payment config data
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/service.CreatePaymentConfigRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PaymentConfigResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Create payment config'
      tags:
      - Admin-Payment-Management
  /admin/payments/configs/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a payment configuration
      parameters:
      - description: Payment config ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.StandardResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Delete payment config'
      tags:
      - Admin-Payment-Management
    put:
      consumes:
      - application/json
      description: Update a payment configuration
      parameters:
      - description: Payment config ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated payment config data
        in: body
        name: config
        required: true
        schema:
          $ref: '#/definitions/service.UpdatePaymentConfigRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PaymentConfigResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[Admin] Update payment config'
      tags:
      - Admin-Payment-Management
...
  /payments/configs:
    get:
      consumes:
      - application/json
      description: Get active payment configurations for public display
      parameters:
      - description: Filter by currency
        example: '"CNY"'
        in: query
        name: currency
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PaymentConfigResponse'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      summary: '[Public] Get active payment configs'
      tags:
      - User-Payment
  /payments/methods:
    get:
      consumes:
      - application/json
      description: Get available payment methods grouped by gateway
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  additionalProperties:
                    items:
                      type: string
                    type: array
                  type: object
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      summary: '[Public] Get available payment methods'
      tags:
      - User-Payment
  /payments/notify/{gateway}:
    post:
      consumes:
      - application/json
      description: Handle payment notification from gateway
      parameters:
      - description: Payment gateway
        enum:
        - epay
        - epusdt
        in: path
        name: gateway
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            type: string
        "400":
          description: fail
          schema:
            type: string
        "500":
          description: fail
          schema:
            type: string
      summary: '[Webhook] Payment notification'
      tags:
      - User-Payment
  /payments/orders:
    post:
      consumes:
      - application/json
      description: Create a new payment order
      parameters:
      - description: Payment order data
        in: body
        name: payment_order
        required: true
        schema:
          $ref: '#/definitions/handler.CreatePaymentOrderRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PaymentRecordResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Create payment order'
      tags:
      - User-Payment
  /payments/orders/{payment_no}:
    get:
      consumes:
      - application/json
      description: Get payment order details
      parameters:
      - description: Payment number
        in: path
        name: payment_no
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.StandardResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.PaymentRecordResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.BadRequestResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.ForbiddenResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.NotFoundResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Get payment order'
      tags:
      - User-Payment
  /payments/orders/my:
    get:
      consumes:
      - application/json
      description: Get current user's payment orders
      parameters:
      - description: Limit results
        example: 10
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - description: Offset results
        example: 0
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.PaymentRecordResponse'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.UnauthorizedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.InternalServerErrorResponse'
      security:
      - BearerAuth: []
      summary: '[User] Get my payment orders'
      tags:
      - User-Payment
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"